DRAFT: Port MONO_IOMAP to CoreFX.
The original mono-io-portability code mapped almost 1:1 with the CoreFX I/O
implementation with the exception of g_dir_open, which for non-Win32 is just
opendir with more memory allocations. The tabs/style are fixed in the function
implementations, while the original portability code is preserved in its
entirety, hideous whitespace and all.

The functions that use a 'const char* path' but were NOT present in
mono-io-portability include the following:
- FnMatch
- ReadLink
- INotifyAddWatch
- RealPath
- LChflags

Peculiarly, Mono still seems to use the old MonoIO _just_ enough to where we
can't name this something more appropriate like DOTNET_IOMAP, so instead we
have to share the old name to get full coverage for this feature.